+++ oshi-core/src/test/java/oshi/driver/linux/proc/UserGroupInfoTest.java
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2021 The OSHI Project Contributors: https://github.com/oshi/oshi/graphs/contributors
+ * Copyright (c) 2021-2022 The OSHI Project Contributors: https://github.com/oshi/oshi/graphs/contributors
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -31,6 +31,7 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import oshi.util.ExecutingCommand;
+import oshi.util.UserGroupInfo;
 
 class UserGroupInfoTest {