+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 6.0.1 (in progress)
 
-* Your contribution here!
+##### Bug fixes / Improvements
+* [#1831](https://github.com/oshi/oshi/pull/1831): Improve Solaris and AIX process listing using procfs - [@dbwiddis](https://github.com/dbwiddis).
 
 # 6.0.0 (2021-12-31)